// crypto-rn/frontend/src/generated/ScreenerListRefetchableQuery.graphql.ts
export type AssetSortInput = {
  name?: SortEnumType | null;
  price?: AssetPriceSortInput | null;
  slug?: SortEnumType | null;
  symbol?: SortEnumType | null;
};
// ...
export type AssetFilterInput = {
  and?: ReadonlyArray<AssetFilterInput> | null;
  name?: StringOperationFilterInput | null;
  or?: ReadonlyArray<AssetFilterInput> | null;
  price?: AssetPriceFilterInput | null;
  slug?: StringOperationFilterInput | null;
  symbol?: StringOperationFilterInput | null;
};
// ...
export type ScreenerListRefetchableQuery$variables = {
  count?: number | null;
  cursor?: string | null;
  order?: ReadonlyArray<AssetSortInput> | null;
  where?: AssetFilterInput | null;
};
export type ScreenerListRefetchableQuery$data = {
  readonly " $fragmentSpreads": FragmentRefs<"screenerListFragment_query">;
};
export type ScreenerListRefetchableQuery = {
  response: ScreenerListRefetchableQuery$data;
  variables: ScreenerListRefetchableQuery$variables;
};